Moved IntTreeMap and *ArrayIterable to collections
