+++ core/src/main/java/org/mini2Dx/core/collections/IntArrayIterable.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package org.mini2Dx.core.util;
+package org.mini2Dx.core.collections;
 
 import org.mini2Dx.gdx.utils.IntArray;
