+++ ui/src/main/java/org/mini2Dx/ui/render/ParentRenderNode.java
@@ -18,9 +18,8 @@ package org.mini2Dx.ui.render;
 import org.mini2Dx.core.Mdx;
 import org.mini2Dx.core.geom.Rectangle;
 import org.mini2Dx.core.Graphics;
-import org.mini2Dx.core.util.IntTreeMap;
+import org.mini2Dx.core.collections.IntTreeMap;
 import org.mini2Dx.gdx.utils.IntMap;
-import org.mini2Dx.ui.UiContainer;
 import org.mini2Dx.ui.element.ParentUiElement;
 import org.mini2Dx.ui.layout.FlexLayoutRuleset;
 import org.mini2Dx.ui.layout.ImmediateLayoutRuleset;
